Prevent cutting stop bit short on new byte
There is a small race condition that occurs when the stop bit of the
last byte in the buffer is transmitted. Instead of waiting for the stop
bit to complete, tx_state is set to 0 and further interrupts are
disabled. If, during the stop bit, a new byte is written, its start bit
starts after 16 timer ticks, without waiting for the stop bit to
complete.

This can be easily reproduced like this:

  altSerial.begin(9600);
  altSerial.write('X');
  delayMicroseconds(925);
  altSerial.write('X');

To fix this, two things are changed:
 - If the buffer is empty and the stop bit just started (tx_state == 10), an
   interrupt is scheduled for the end of the stop bit (setting tx_state = 11).
 - If the buffer is empty and the stop bit is done (tx_state == 11), finish
   sending (sending tx_state = 0).
 - If there are more bytes and the stop bit has just started (tx_state == 10),
   an interrupt is scheduled for the end of the stop bit, which also clears the
   line to send the start bit.
 - If there are more bytes and the stop bit has just finished (tx_state == 11),
   an interrupt is scheduled in 16 timer ticks, which will also clears the line
   to send the start bit. Note that this re-queries the current timer
   count, when using target + 16 it is likely that that time has already
   passed.

The last point should not occur often, this is exactly the old race condition.
It might seem easier to handle the last case in writeByte(), by checking
for tx_state == 11 and doing CONFIG_MATCH_CLEAR() to let the next
compare match send the start bit. However, this introduces a new race
condition where the compare match could happen between checking tx_state
and doing CONFIG_MATCH_CLEAR(), missing the window. Additional checks
could detect this, but then the code would probably be more complicated
again.