Simplify TX interrupt handler a bit
This merges the handling of the stop bit into the loop for data bits,
which prevents one unneeded interrupt, when the last data bit sent
already has the TX line high.

In this case, there is no need for an interrupt at the begin of the stop
bit and instead the next interrupt can happen at the begin of the next
byte's start bit (or at the end of the stop bit if there are no more
bytes queued).

As an added bonus, this change saves 6 bytes of compiled code.